requirements.txt
altgraph==0.16.1
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
constantly==15.1.0
enum34==1.1.6
future==0.17.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
macholib==1.11
numpy==1.16.4
paho-mqtt==1.4.0
pefile==2019.4.18
Pillow==6.1.0
PyGetWindow==0.0.6
PyHamcrest==1.9.0
PyMsgBox==1.0.7
pynput==1.4.2
PyRect==0.1.4
PyScreeze==0.1.22
pyTelegramBotAPI==3.6.6
PyTweening==1.0.3
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.25.3
zope.interface==4.6.0
